Created a method to calculate the average of given number of stocks and purchased prices and also create another method for calculating the total number of stocks just for the viewing purposes. Created a button using for loop so that user can press to create more fields to enter.
The average and total amount of stocks will not be visible until "calculate" button has been pressed.
Total number of stocks limited to 2 decimal points, due to possibilities of fractional buys
Average amount limited to 2 decimal points for visibility

Created FocusState for hiding the keypad when the user taps elsewhere from the textfield.